Refactoring.
Keeping the transactions abstract, to allow for simulations that either model
the UTxO, or abstract from it.

Also, added the concept of a delayed computation, to represent local
computations that take some time. We had this for transaction validation, with a
manual thread delay, but we will encounter this more often as we implement the
full protocol.