Add ability for nodes to submit multiple transactions.
The naive way of submitting them all at once is both a bit boring (because they
will all end up in the same snapshot), and dangerousm since it is prone to
overflowing queues. So we added monitoring the set of transactions in flight,
and sending new transactions only after older ones have been confirmed.