Fix: don't reenqueue to the front of the queue.
When we were handling a message that was not yet valid, we were immediately
reenqueueing it. In the case that the queue did not have any other messages,
this would lead to an infinite loop.

Instead, we defer reenqueueing the event until there is at least one other event
in the queue.