Changes to the Channel class.
Instead of having 'recv :: m (Maybe a)', we have it be an STM action, 'recv ::
STM m a', to make it composable.

Also, with the multiplexer, the 'Channel' will only be responsible for the delay
during transport (not the serialisation part, which will be modelled on the
sending and receiving side instead), so we change the delayed channel to exclude
the size dependent part.