Introduce Multiplexer, to better model message delay.
Note: this implements the multiplexer, but does not use it yet.

We will model the timing of messages being sent between nodes in a network.

The network itself is modelled by a delay in time between each message leaving
the sending node and its arrival at the target node. The delay is determined by
the distance between nodes, and is constant for any pair of nodes. In
particular, it does not depend on the message size.

Before being sent across the network, each message has to be serialised. So the
event of a message being sent (or being received) by a node does not correspond
to a single point in time, but rather to a time interval. We take account for
that by modelling each message by its /leading/ and /trailing edge/ (see
'MessageEdge'). The time distance between leading and trailing edge of a message
will typically be linear in the message size (though we allow supplying any
function from message size to length in time, see 'mpWriteCapacity').

In addition to delaying messages through serialisation and network transport, we
also model two points of contention: on the egress side, nodes from one node to
all its peers have to pass through one networking interface, which supports a
finite data rate. Similarly, on the ingress side, arriving messages from all
nodes have to pass through one networking interface.

We model sending and receiving a message as follows:

- The sending node has one queue for all outgoing messages. An outgoing message
  is placed in this queue, together with the target 'NodeId'.

- There is one thread that reads messages from the queue. When it reads a
  message, it will split it into its leading and trailing edge.

    1. It will immediately put the leading edge in a 'Channel' to the target node

    2. It will sleep for a time given my 'mpWriteCapacity' applied to the message size.

    3. It will then put the trailing edge of the message into the queue.

  This models the contention and serialisation delay. Note that because there is
  only one thread sending messages, there is no need for explicit locking.

- Each 'Channel' will have a specified delay between a message being sent and
  received.

  This allows us to model the network delay due to geographical distribution.

- On the receiving side, we again model contention by having one thread that
  listens to all incoming channels.

    1. When it sees the leading edge of a message, it sleeps for a time given my
    'mpReadCapacity' applied to the message size.

    2. When it sees the trailing edge of a message, it puts the message itself
    into the inbox of the node (modelled as a queue holding messages paired with
    the 'NodeId' of the sending node).

Possible Improvements:

- Add some variance to the network delay.

- If messages do not fit the current TCP window, the network delay will
  increase. We do not model this yet.