Performance optimisation.
It is safe to introduce a little bit of parallelism to the nodes: after
receiving a message/snapshot to sign, they first validate it against their local
state, and then, if valid, sign it and send a reply. Since signing and sending
does not depend on the local state, we can perform this in a separate thread.